Fix x64 version doesn't compile.
git-svn-id: https://subversion.assembla.com/svn/q3/trunk@3477 8af8166b-12a6-a448-a533-9086ace3f9f6

+++ q3/qmsmtp/src/ui.cpp
@@ -190,7 +190,7 @@ LRESULT qmsmtp::PopBeforeSmtpDialog::onInitDialog(HWND hwndFocus,
 		ComboBox_AddString(hwndProtocol, ptszName);
 		
 		if (wcscmp(pData_->wstrProtocol_.get(), pwszName) == 0)
-			nSelection = n;
+			nSelection = static_cast<int>(n);
 	}
 	ComboBox_SetCurSel(hwndProtocol, nSelection);
